Fix MQTT session expiry persistence

+++ src/main/java/io/mapsmessaging/engine/session/SessionManagerPipeLine.java
@@ -142,7 +142,7 @@ public class SessionManagerPipeLine {
   void close(SessionImpl sessionImpl, boolean clearWillTask) {
     SubscriptionController subscriptionController;
     long expiry = sessionImpl.getExpiry();
-    sessions.remove(sessionImpl.getName());
+    sessions.remove(sessionImpl.getName(), sessionImpl);
     String storeName = (sessionImpl instanceof PersistentSession) ?  ((PersistentSession)sessionImpl).getStoreName(): "";
     sessionImpl.close();
 
@@ -189,8 +189,11 @@ public class SessionManagerPipeLine {
 
   void closeAndDeleteSubscriptionController(String sessionStateFile, SubscriptionController subscriptionController) {
     closeSubscriptionController(subscriptionController);
+    storeLookup.removeSessionDetails(subscriptionController.getSessionId(), subscriptionController.getUniqueSessionId());
     try {
-      Files.deleteIfExists(new File(sessionStateFile).toPath());
+      if (!sessionStateFile.isEmpty()) {
+        Files.deleteIfExists(new File(sessionStateFile).toPath());
+      }
     } catch (IOException e) {
       // ignore
     }
@@ -200,7 +203,7 @@ public class SessionManagerPipeLine {
     if (subscriptionController.getTimeout() != null) {
       expiredSessions.increment();
     }
-    subscriptionManagerFactory.remove(subscriptionController.getSessionId());
+    subscriptionManagerFactory.remove(subscriptionController.getSessionId(), subscriptionController);
     WillTaskImpl willTaskImpl = willTaskManager.remove(subscriptionController.getSessionId());
     if (willTaskImpl != null) {
       willTaskImpl.cancel();
@@ -219,7 +222,7 @@ public class SessionManagerPipeLine {
   //
   private SubscriptionController loadSubscriptionManager(SessionContext context) {
     context.setRestored(false);
-    SessionDetails sessionDetails = storeLookup.getSessionDetails(context);
+    SessionDetails sessionDetails = context.isResetState() ? storeLookup.resetSessionDetails(context) : storeLookup.getSessionDetails(context);
     context.setUniqueId(sessionDetails.getUniqueId());
     context.setInternalSessionId(sessionDetails.getInternalUnqueId());
     SubscriptionController subscriptionManager = subscriptionManagerFactory.get(context.getId());
@@ -239,7 +242,7 @@ public class SessionManagerPipeLine {
       if (context.isResetState()) {
         disconnectedSessions.decrement(); // No longer stored
         logger.log(ServerLogMessages.SESSION_MANAGER_FOUND_EXISTING, context.getId(), context.isResetState());
-        subscriptionManagerFactory.remove(context.getId());
+        subscriptionManagerFactory.remove(context.getId(), subscriptionManager);
         subscriptionManager.close(false);
         sessionDetails.clearSubscriptions();
         subscriptionManager = new SubscriptionController(context, destinationManager, new LinkedHashMap<>());
